requirements.txt
openai>=1.2.0

pyautogui
pillow
python-dotenv
